Fixes activity markers
Prevents activity markers from showing when the map is zoomed out.

Note: there may be a better fix for this issue.

When zoomed in and the map is refreshed, all the markers for the
current day are created (pending result from Glide), then all
the city markers are created (also pending a result from Glide).

However, the Glide callback for the current day may not be invoked
until after the markers array has been cleared in preparation for
adding city markers.

This results in activity markers can be drawn in over top of the
city marker.